Moved DTOs to different directories
