+++ Start/app/src/main/java/com/esei/grvidal/nighttime/network/network_DTOs/CalendarDTO.kt
@@ -1,0 +1,10 @@
+package com.esei.grvidal.nighttime.network.network_DTOs
+
+import com.esei.grvidal.nighttime.network.EventData
+
+data class CalendarData(
+    val total: Int,
+    val friends: Int,
+    val events: List<EventData>
+)
+
